Update signalProcessing2C.py
Pared down version of main script to troubleshoot saving spatial distribution plots. These plots would show where in an image various measurements were recorded. E.g., the period measurements should be pretty homogenous across the field of view, but they may not be if the wave character is different in different parts of the FOV and this would be interesting to see at a glance.