Upload wheels
## Summary
Adds a CI workflow that uploads wheels to pypi for pushed tags. This means that users should be able to run the application without cloning the whole repository. These changes are in pursuit of https://github.com/zacswider/waveAnalysis/issues/18, although adding the script file will be a subsequent PR.

## Major Changes
 * Adds a CI file that runs pytest on all pushes and uploads wheels to pypi on tagged pushes
 * Pins specific version numbers of all dependencies
 * Updates tests to assert all close instead of all exact because of some tiny rounding differences between test data and runtime data created with different package versions.

## Minor Changes
 * formatting fixes on a handful of files
 * Specifies hatchling as the build backend
 * Adds a `uv.lock` file (see https://docs.astral.sh/uv/concepts/projects/layout/#the-lockfile)